Adding entities and Services

+++ finance-mgr-backend/src/main/java/com/soumen/app/finance_mgr_backend/data/entities/BankAccount.java
@@ -21,7 +21,7 @@ public abstract class BankAccount {
     @ManyToOne
     @JoinColumn(name = "bank_id")
     private Bank bank;
-    private BigDecimal balance;
+    private Double balance;
     private AccountType accountType;
     private String accountNumber;
     private Double rateOfInterest;
